refactor: migrate codebase from Java to Kotlin
