+++ src/main/java/com/group/libraryapp/domain/user/loanhistory/JavaUserLoanHistory.java
@@ -11,7 +11,7 @@ import org.jetbrains.annotations.NotNull;
 import static javax.persistence.GenerationType.IDENTITY;
 
 @Entity
-public class UserLoanHistory {
+public class JavaUserLoanHistory {
 
   @Id
   @GeneratedValue(strategy = IDENTITY)
@@ -24,11 +24,11 @@ public class UserLoanHistory {
 
   private boolean isReturn;
 
-  public UserLoanHistory() {
+  public JavaUserLoanHistory() {
 
   }
 
-  public UserLoanHistory(User user, String bookName, boolean isReturn) {
+  public JavaUserLoanHistory(User user, String bookName, boolean isReturn) {
     this.user = user;
     this.bookName = bookName;
     this.isReturn = isReturn;
