+++ src/main/java/com/group/libraryapp/domain/user/loanhistory/UserLoanHistoryRepository.java
@@ -1,5 +1,6 @@
 package com.group.libraryapp.domain.user.loanhistory;
 
+import com.group.libraryapp.domain.UserLoanHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserLoanHistoryRepository extends JpaRepository<UserLoanHistory, Long> {
