+++ src/main/kotlin/com/group/libraryapp/domain/UserLoanHistory.kt
@@ -1,0 +1,27 @@
+package com.group.libraryapp.domain
+
+import com.group.libraryapp.domain.user.User
+import javax.persistence.Entity
+import javax.persistence.GeneratedValue
+import javax.persistence.GenerationType
+import javax.persistence.Id
+import javax.persistence.ManyToOne
+
+@Entity
+class UserLoanHistory(
+    @ManyToOne
+    val user: User,
+
+    val bookName: String,
+
+    var isReturn: Boolean,
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    val id: Long? = null,
+) {
+
+    fun doReturn() {
+        this.isReturn = true
+    }
+}
